refactor(workers): extract consumer runtime
